gql_plannedlessons/requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
fastapi[all]==0.70.0
uvicorn[standard]==0.13.4
gunicorn==20.0.4
inotify==0.2.10

psycopg2-binary==2.9.1
asyncpg==0.26.0

sqlalchemy==1.4.26
sqlalchemy_utils==0.37.9

pydantic==1.8.2
starlette==0.16.0
strawberry-graphql==0.119.0


https://github.com/hrbolek/uoishelpers/archive/refs/heads/main.zip